* Fixed alphabetical sort for games with longer game titles * Redid DVDSelectGame function to use cDI magic to select the file format instead of trying to open the files and see if the game exists in that format * Added offset correction for dl games (at least for Metroid other M and Xenoblade Chronicles). Those game now load the health warning screen in an endless loop and still not working. This is a work in progress

+++ di/dip.h
@@ -172,7 +172,7 @@ typedef struct
 
 typedef struct
 {
-	u32 Offset;
+	u64 Offset;
 	u32 Size;
 	s32 File;
 } FileCache;
